Removed 2nd eeprom logic and USB logic since we don't plan on populating either of those moving forward.
For now the driver presumes that all analog sensors are present if the ADC is present. I need to set these to the default for the fully populated board we have for now. We can add board revision logic later if necessary.

For now the driver presumes that the motion sensor is present - there's no I2C for it so we can't detect it AFAIK. We need to get board revision switch in for this.