omap3isp: Don't miss the first frame when starting the video stream
The ISP pipeline is enabled before starting the video buffers queue.
This results in video::set_stream() being called on the output module
without any buffer being queued on its output. As at least the resizer
doesn't handle being started in the middle of a frame correctly, the
last module in the chain is then started on the first frame
synchronization interrupt. This results in the first frame being lost.

Swap the isp_pipeline_set_stream() and isp_video_queue_streamon() in
isp_video_streamon() so that the last module in the pipeline will have
an output buffer when being started. Change the CSI2, CCDC, preview and
resizer modules to handle the .buffer_queue callback when the stream is
stopped, and start the modules synchronously on video::set_stream() if a
buffer is available at the output.

Signed-off-by: Laurent Pinchart <laurent.pinchart@ideasonboard.com>